Se refactoriza cambiando BigDecimal por Double en campos de modelos y entidades

+++ domain/model/src/main/java/co/com/crediya/model/loantype/LoanType.java
@@ -18,7 +18,7 @@ public class LoanType {
     String code;
     Double minimumAmount;
     Double maximumAmount;
-    BigDecimal interestRate;
+    Double interestRate;
     Boolean automaticValidation;
 
 }
